requirements.txt
openai>=1.40.0
requests>=2.31.0
python-dateutil>=2.9.0
pytz>=2024.1
gspread>=6.0.0
google-auth>=2.34.0
discord.py>=2.4.0
